tests/gpu/compatible_mode/test_model/huggingface/inference/requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autocommand==2.2.2
av==12.3.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
boto3==1.34.149
botocore==1.34.149
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
cloudpickle==3.0.0
datasets==2.20.0
diffusers==0.29.2
dill==0.3.8
distro==1.9.0
evaluate==0.4.2
exceptiongroup==1.2.2
filelock==3.15.4
flake8==7.1.1
frozenlist==1.4.1
fsspec==2024.5.0
fvcore==0.1.5.post20221221
huggingface-hub==0.24.2
hypothesis==6.108.5
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
iniconfig==2.0.0
invisible-watermark==0.2.0
iopath==0.1.10
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
lark==1.1.9
lazy_loader==0.4
libcst==1.4.0
MarkupSafe==2.1.5
mccabe==0.7.0
MonkeyType==23.3.0
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx
numpy==1.23.5
opencv-python==4.10.0.84
optree==0.12.1
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
patch==1.16
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.1
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
pynvml==11.5.3
pytest==8.3.2
pytest-benchmark==4.0.0
python-dateutil==2.9.0.post0
pytorchvideo==0.1.5
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn
scipy
sentence-transformers==3.0.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
submitit==1.5.1
sympy==1.13.1
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.24
timm
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
transformers==4.44.0
triton==3.0.0
typeguard==4.3.0
types-dataclasses==0.6.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
